Crate ecs_client

Reexports

extern crate std as std;

extern crate hyper;

use std::prelude::v1::*;

Modules

ecs_client
This module contains the ECSClient which can be used to interact with Amazon ECS's API.

params
Defines structs which contain the relevant parameters for each type of request to Amazon ECS. Each can be serialized to a json blob and set as the body of the HTTP request.

region
An enum defining the regions in which Amazon ECS is supported.

request
A Signature Version 4 Signable AWS Request to ECS