Module ecs_client::action::ecs_action

This module contains types which are useful for describing actions that the ECSClient can perform on your behalf.

Reexports

use std::prelude::v1::*;

use std::fmt::{Display, Formatter, Error};

use serde::ser::Serialize;

use serde::de::Deserialize;

Enums

ECSAction
An enum defining the possible actions which can be sent to Amazon ECS.

Traits

ECSRequest
A marker trait to "supertype" each particular kind of ECS request. Each ECSRequest must be serde::ser::Serialize so that it can be serialzied to json and set as the payload (body) of an HTTP request.

ECSResponse
A marker trait to "supertype" each particular kind of ECS response. Each ECSResponse must be serde::de::Deserialize so that it can be deserialized from the json blob in the body of the HTTP response.