Module ecs_client::action::list_clusters

Defines request and response types for a ListClusters action.

Reexports

use std::prelude::v1::*;

use action::ecs_action::{ECSRequest, ECSResponse};

Structs

ListClustersRequest
A ListClusters request type which can be serialized to json and set as the body of an HTTP request. Construct one of these and pass it to your client in the ecs_client::list_clusters() function.

ListClustersResponse
A ListClusters response type which can be deserialized from the body of an HTTP response. If the request was successful, one of these will be returned from ecs_client::list_clusters().