Struct ecs_client::action::list_clusters::ListClustersResponse

pub struct ListClustersResponse {
    clusterArns: Vec<String>,
    nextToken: Option<String>,
}

A ListClusters response type which can be deserialized from the body of an HTTP response. If the request was successful, one of these will be returned from ecs_client::list_clusters().

Fields

clusterArns

The list of full Amazon Resource Name (ARN) entries for each cluster associated with your account.

nextToken

A value returned by the previous request indicating where to begin the next page of paginated output. This value will be None if there are no more results to return. Otherwise, this value will hold Some(String) which can be used in a subsequent ListClustersRequest to obtain the next page of paginated output, where the previous output exceeded 100 results or the value indicated by maxResults in the ListClustersRequest.

Methods

impl ListClustersResponse

Implements some convenience methods for looking at values returned in a ListClustersResponse.

fn get_cluster_arns(&self) -> &Vec<String>

Gets a reference to the Vec of Amazon Resource Name (ARN) Strings describing your Amazon ECS Compute Clusters.

fn get_next_token(&self) -> &Option<String>

Gets a reference to the nextToken value returned by the previous request. This will return &None of there are no more results to display, otherwise it will return Some(String) which can be used in a subsequent request to get the next page of paginated output.