Struct ecs_client::ecs_client::ECSClient

pub struct ECSClient {
    region: Region,
    client: Client,
}

Fields

region

client

Methods

impl ECSClient

fn for_region(region: Region) -> ECSClient

Creates a new ECSClient for the specified Region.

fn set_region(&mut self, region: Region)

Sets the Region to which the client sends requests.

fn list_clusters(&self, request: ListClustersRequest) -> ListClustersResponse

Lists all of your compute clusters on ECS.

fn sign_and_send<T: ECSRequest>(&self, action: ECSAction, request: T) -> Response

Creates an HTTP request to be sent to Amazon ECS. Signs the request using Amazon's Signature Version 4 Signing Algorithm. Serializes the service request to json format and sets it as the payload in the HTTP body. Sends the request to ECS and returns the hyper::client::Response.

fn build_request_uri(&self) -> String

Builds the request URI based on the Region this client is currently configured to send requests to.

fn build_headers(&self, action: ECSAction, content_length: u64) -> Headers

Builds a hyper::header::Headers with the Host, Accept-Encoding, X-Amz-Target, X-Amz-Date, Content-Type, and Content-Length HTTP headers set.

fn build_hostname(&self) -> String

Builds and returns the hostname String used in the Host header.

fn build_x_amz_target(&self, action: ECSAction) -> String

Builds and returns the target String used in the X-Amz-Target header.