Module ecs_client::request

This module defines request types which can be serialized to json and set as the payload of an HTTP request body.

Reexports

use std::prelude::v1::*;

Modules

ecs_request
A marker trait to "supertype" each particular kind of ECS request. Each ECSRequest must be serde::ser::Serialize so that it can be serialized to json and set as the payload of the request in the HTTP body.

list_clusters
A struct for a ListClusters request which can be serialized to json and set as the body of an HTTP request.