Function ecs_client::signature::build_string_to_sign

fn build_string_to_sign(headers: &Headers, region: Region, serv_abbrev: &str, hashed_canon_req: &str) -> (String, String)

Builds the String To Sign according to the guidelines at http://docs.aws.amazon.com/general/latest/gr/sigv4-create-string-to-sign.html . The String To Sign is used with a derived signing key to calculate the signature. Returns a tuple containing (string_to_sign, credential_scope) so that credential scope value can be reused in the Authorization HTTP header.