Function ecs_client::signature::derive_signing_key

fn derive_signing_key(headers: &Headers, region: Region, serv_abbrev: &str) -> [u8; 32]

Derives the signing key from your AWS secret access key, the date of your request, the service name, and the region the request is being sent to. AWS credentials are sourced from environment variables. Your AWS secret access key must be stored in an environment variable called AWS_SECRET_ACCESS_KEY.