Crate ecs_client

This crate contains an ECSClient which can be used to interact with Amazon ECS's API. The client sources your AWS credentials from the environment in which it is running. You'll need to create an AWS IAM user and add them to a security group with proper permissions, as outlined here: http://docs.aws.amazon.com/AmazonECS/latest/developerguide/get-set-up-for-amazon-ecs.html . Once you've created an IAM user, you can generate security credentials by going to the IAM console, clicking on Users in the navigation pane, selecting your user from the list, choosing the Security Credentials tab, and clicking Create Access Key. This will generate an AWS access key ID and an AWS secret access key. Store these in the environment in which your client will be running by setting the AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY, respectively, as such:

$ export AWS_ACCESS_KEY_ID="your_access_key_id"
$ export AWS_SECRET_ACCESS_KEY="your_secret_access_key"

With these credentials created and stored appropriately, the ECSClient should be able to make requests to Amazon ECS on your behalf. Simply create an ECSClient, build a request, and call the relevent API action on the client. For example (note that the code below should panic if no security credentials have been stored in the environment):

extern crate ecs_client;

use ecs_client::ecs_client::ECSClient;
use ecs_client::region::Region;
use ecs_client::action::*;

fn main() {
    let client = ECSClient::for_region(Region::USWest2);
    let request = list_clusters::ListClustersRequest::new();
    let response: list_clusters::ListClustersResponse = ecs_client.list_clusters(request);
    for cluster_arn in response.get_cluster_arns() {
        println!("{}", cluster_arn);
    }
}

Reexports

extern crate std as std;

extern crate serde;

extern crate serde_json;

extern crate hyper;

extern crate time;

extern crate sodiumoxide;

use std::prelude::v1::*;

Modules

action
This module defines request and response types for the various actions supported by ECS.

custom_headers
Defines the X-Amz-Target and X-Amz-Date custom HTTP headers using hyper macros

custom_ser
This module contains functions called by serde to customize serialization of ECSRequest types.

ecs_client
This module contains the ECSClient which can be used to make API requests to Amazon ECS.

error
This module defines Error types that may be returned when sending a request to Amazon ECS. Note that the ECS API reference defines 'Exception' types, but we map each Amazon ECS Exception to a Rust Error.

region
An enum defining the regions in which Amazon ECS is supported.

signature
This module implements Amazon's Signature Version 4 Signing Algorithm which is required in order to make authenticated requests to any Amazon Web Service.