Crate ecs_client

Reexports

extern crate std as std;

extern crate serde;

extern crate serde_json;

extern crate hyper;

extern crate time;

extern crate sodiumoxide;

use std::prelude::v1::*;

Modules

action
This module defines request and response types for the various actions supported by ECS.

custom_headers
Defines the X-Amz-Target and X-Amz-Date custom HTTP headers using hyper macros

ecs_client
This module contains the ECSClient which can be used to interact with Amazon ECS's API.

error
This module defines Error types that may be returned when sending a request to Amazon ECS. Note that the ECS API reference defines 'Exception' types, but we map each Amazon ECS Exception to a Rust Error.

region
An enum defining the regions in which Amazon ECS is supported.

signature
This module implements Amazon's Signature Version 4 Signing Algorithm which is required in order to make authenticated requests to any Amazon Web Service.